[FIX] website_slides@saas~12.2: define column with correct type

+++ migrations/website_slides/saas~12.2.2.0/pre-models.py
@@ -86,7 +86,7 @@ def migrate(cr, version):
             channel_id int4,
             partner_id int4,
             completed boolean,
-            completion float8
+            completion int4
         )
     """
     )
